requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
psycopg2-binary==2.9.11
